vkorzykov change in TaskB1 and C1.

+++ src/vkorzykov/Lab2/TaskC1.java
@@ -4,7 +4,7 @@ package vkorzykov.Lab2;
 import java.util.Scanner;
 
 public class TaskC1 {
-    public static double convertCelsiumToForenheit (float celsium) {
+    public static double convertCelsiumToForenheit (int celsium) {
         return celsium * 9 / 5 + 32;
     }
